requirements.txt
amqp==2.5.2
asn1crypto==0.24.0
django==2.2.2
djangorestframework
attrs==19.3.0
backcall==0.1.0
billiard==3.6.1.0
bleach==3.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
cryptography==2.1.4
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
idna==2.8
importlib-metadata==1.2.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
keyring==10.6.0
keyrings.alt==3.0
Kivy==1.11.1
Kivy-Garden==0.1.4
kombu==4.6.6
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
peewee==3.13.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.26.1
pymongo==3.9.0
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
pyzmq==18.1.1
qtconsole==4.6.0
redis==3.3.11
requests==2.22.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sh==1.12.14
six==1.11.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.8
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
gevent==1.4.0
eventlet==0.25.1